Update gif250 to gro525
